Moved util.py into utils.http package
